+++ huanshankeji-team/gradle-library/src/main/kotlin/com/huanshankeji/team/gitversioning/gitlabpackageregistry/Repository.kt
@@ -3,57 +3,71 @@ package com.huanshankeji.team.gitversioning.gitlabpackageregistry
 import com.huanshankeji.GradleCommonExperimentalApi
 import com.huanshankeji.artifacts.MavenRepositoryHandlerContext
 import com.huanshankeji.artifacts.leadingProjectNameModuleRegex
-import com.huanshankeji.gitlab.packageregistry.maven.GITLAB_COM_HOST
+import com.huanshankeji.gitlab.packageregistry.maven.gitlabPackageRegistryProjectLevelEndpointMavenRepository
 import com.huanshankeji.gitversioning.ConventionVersionRegexes
 import com.huanshankeji.gitversioning.gitlabpackageregistry.gitlabPackageRegistryProjectLevelEndpointConventionMavenRepositories
 import org.gradle.api.artifacts.dsl.RepositoryHandler
 import org.gradle.api.provider.ProviderFactory
 import java.net.URI
 
+/**
+ * @param projectIdOrProjectPath See [RepositoryHandler.gitlabPackageRegistryProjectLevelEndpointMavenRepository].
+ */
 context(providers: ProviderFactory, _: (path: Any) -> URI)
 fun RepositoryHandler.huanshankejiGitlabPackageRegistryProjectEndpointConventionMavenRepositories(
-    projectId: String,
+    projectIdOrProjectPath: String,
     groupRegex: String,
     moduleRegex: String,
     versionRegexes: ConventionVersionRegexes = ConventionVersionRegexes(),
 ) =
     gitlabPackageRegistryProjectLevelEndpointConventionMavenRepositories(
-        GITLAB_COM_HOST, projectId, groupRegex, moduleRegex, versionRegexes
+        projectIdOrProjectPath = projectIdOrProjectPath,
+        groupRegex = groupRegex,
+        moduleRegex = moduleRegex,
+        versionRegexes = versionRegexes,
     )
 
 @GradleCommonExperimentalApi
 fun MavenRepositoryHandlerContext.huanshankejiGitlabPackageRegistryProjectEndpointConventionMavenRepositories(
-    projectId: String,
+    projectIdOrProjectPath: String,
     groupRegex: String,
     moduleRegex: String,
     versionRegexes: ConventionVersionRegexes = ConventionVersionRegexes(),
 ) =
     context(providers, uri) {
         repositories.huanshankejiGitlabPackageRegistryProjectEndpointConventionMavenRepositories(
-            projectId, groupRegex, moduleRegex, versionRegexes
+            projectIdOrProjectPath, groupRegex, moduleRegex, versionRegexes
         )
     }
 
+/**
+ * @param projectIdOrProjectPath See [RepositoryHandler.gitlabPackageRegistryProjectLevelEndpointMavenRepository].
+ * @param projectName serves as the prefix for the default [leadingProjectNameModuleRegex] `moduleRegex`.
+ */
 context(providers: ProviderFactory, _: (path: Any) -> URI)
 fun RepositoryHandler.defaultHuanshankejiGitlabPackageRegistryProjectEndpointConventionMavenRepositories(
-    projectId: String,
-    group: String,
+    projectIdOrProjectPath: String,
     projectName: String,
+    group: String,
     versionRegexes: ConventionVersionRegexes = ConventionVersionRegexes(),
 ) =
     huanshankejiGitlabPackageRegistryProjectEndpointConventionMavenRepositories(
-        projectId, Regex.escape(group), leadingProjectNameModuleRegex(projectName), versionRegexes
+        projectIdOrProjectPath, Regex.escape(group), leadingProjectNameModuleRegex(projectName), versionRegexes
     )
 
+/**
+ * @param projectIdOrProjectPath See [RepositoryHandler.gitlabPackageRegistryProjectLevelEndpointMavenRepository].
+ * @param projectName serves as the prefix for the default [leadingProjectNameModuleRegex] `moduleRegex`.
+ */
 @GradleCommonExperimentalApi
 fun MavenRepositoryHandlerContext.defaultHuanshankejiGitlabPackageRegistryProjectEndpointConventionMavenRepositories(
-    projectId: String,
-    group: String,
+    projectIdOrProjectPath: String,
     projectName: String,
+    group: String,
     versionRegexes: ConventionVersionRegexes = ConventionVersionRegexes(),
 ) =
     context(providers, uri) {
         repositories.defaultHuanshankejiGitlabPackageRegistryProjectEndpointConventionMavenRepositories(
-            projectId, group, projectName, versionRegexes
+            projectIdOrProjectPath, projectName, group, versionRegexes
         )
     }