+++ kotlin-common/gradle-library/src/main/kotlin/com/huanshankeji/artifacts/Module.kt
@@ -1,7 +1,7 @@
 package com.huanshankeji.artifacts
 
 /**
- * A conventional Maven module-id regex for a project name.
+ * A conventional Maven module-id regex with a leading project name.
  */
 fun leadingProjectNameModuleRegex(projectName: String): String =
-    "$projectName.*"
+    """$projectName.*"""