+++ kotlin-common/gradle-library/src/main/kotlin/com/huanshankeji/github/packages/maven/GithubPackagesMavenPublishExtension.kt
@@ -1,0 +1,8 @@
+package com.huanshankeji.github.packages.maven
+
+import org.gradle.api.provider.Property
+
+interface GithubPackagesMavenPublishExtension {
+    val owner: Property<String>
+    val repository: Property<String>
+}